% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codak_model_xgboost.R
\name{codak.xgboost}
\alias{codak.xgboost}
\title{Covariate and direction adaptive knockoff procedure with group structure}
\usage{
codak.xgboost(
  p,
  signy,
  x,
  piargs = list(),
  alphaargs = list(),
  betaargs = list(),
  group = NULL,
  initparam = NULL,
  nfits = 20,
  gamma = 1
)
}
\arguments{
\item{p}{a vector of the p-values.}

\item{signy}{a vector of the signs of test statistics.}

\item{x}{a data frame of the accompanying covariates.}

\item{group}{a numerical vector of group labels. It ranges from 1 to \eqn{g}{g}, where \eqn{g}{g} is the number of groups. It is set to NULL by default.}

\item{initparam}{A matrix. The initial parameter of the EM-algorithm. It is generated from the algorithm described in Appendx B.3 in the paper by default.}

\item{nfits}{An integer. The maximum number of model-fitting. Supposing \eqn{n}{n} is the number of null hypotheses, the model will be refitted every time after accepting \eqn{\left \lceil{n/nfits}\right \rceil}{[n/nfits]} null hypotheses.}

\item{gamma}{A number concatenating the second shape parameter for the beta densities. Default to 1.}

\item{eta}{A number or a set of candidate numbers for model selection. The learning rate used in \code{xgboost} package. Default to 0.3.}

\item{max_depth}{A number or a set of candidate numbers for model selection. The maximum depth of the trees used in \code{xgboost} package. Default to 6.}
}
\value{
A vector containing the q-value of each null hypothesis. The idea of q-values is from Storey (2002).
 The rejection rule of \code{codak_group} is equivalent to thresholding on this value.
 Setting the nominal FDR level as \eqn{\alpha}{\alpha} will reject and only reject all the null hypotheses with
 q-values smaller than or equal to \eqn{\alpha}{\alpha}.
}
\description{
Codak is used to solve multiple hypothesis testing problems accompanied by generic covariate information.
   \code{codak.xgboost} fits parameters by \code{\link[xgboost]{xgboost}} from \code{xgboost} package.
   This function can not only utilize numerical covaraites but also work under situations where null hypotheses are divided into multiple groups and accompanied by different sets of covariates in different groups.
}
\examples{
m <- 4000
mu1 <- -1.5
mu2 <- 3
omega=0.25
alpha=0.1

x1 <- 2*runif(m)-1
x2 <- 2*runif(m)-1
x3 <- sample(c(0,1),m,replace=TRUE,prob=c(0.7,0.3))
pi0=0.6+0.3*(1-omega)*(x1^2+x2^2-2/3)-0.375*omega*x1*x2
pi1=(1-pi0)*(1+x1*x2)/2
pi2=1-pi0-pi1
temp=rep(0,m)
for(j in 1:m){
 temp[j]=sample(c(0,1,2),1,replace=TRUE,prob=c(pi0[j],pi1[j],pi2[j]))
}
y=rnorm(m)+mu1*(temp==1)+mu2*(temp==2)

x=data.frame(x1=x1,x2=x2)
pval=2-2*pnorm(abs(y))
out=codak.xgboost(pval,sign(y),x)
rejresult=(out<=alpha)

fdr=sum((rejresult)&(temp==0))/max(sum(rejresult),1)
power=sum((rejresult)&(temp!=0))/max(sum(temp!=0),1)

}
